+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_analyze.py
@@ -1678,6 +1678,8 @@ class TestAnalyze(TextAnalyticsTest):
             actions=[
                 AnalyzeHealthcareEntitiesAction(
                     model_version="latest",
+                    fhir_version="4.0.1",
+                    document_type="HistoryAndPhysical",
                 )
             ],
             show_stats=True,
@@ -1691,6 +1693,7 @@ class TestAnalyze(TextAnalyticsTest):
                     assert res.error.code == "InvalidDocument"
                 else:
                     assert res.entities
+                    assert res.fhir_bundle
                     assert res.statistics
 
     @TextAnalyticsPreparer()