+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_analyze_async.py
@@ -1782,6 +1782,8 @@ class TestAnalyzeAsync(TextAnalyticsTest):
                 actions=[
                     AnalyzeHealthcareEntitiesAction(
                         model_version="latest",
+                        fhir_version="4.0.1",
+                        document_type="HistoryAndPhysical",
                     )
                 ],
                 show_stats=True,
@@ -1798,6 +1800,7 @@ class TestAnalyzeAsync(TextAnalyticsTest):
                         assert res.error.code == "InvalidDocument"
                     else:
                         assert res.entities
+                        assert res.fhir_bundle
                         assert res.statistics
 
     @TextAnalyticsPreparer()