+++ sdk/textanalytics/azure-ai-textanalytics/CHANGELOG.md
@@ -5,7 +5,10 @@
 ### Features Added
 - Added the Extractive Summarization feature and related models: `ExtractSummaryAction`, `ExtractSummaryResult`, and `SummarySentence`.
   Access the feature through the `begin_analyze_actions` API.
-
+- Added keyword arguments `fhir_version` and `document_type` to `begin_analyze_healthcare_entities` and `AnalyzeHealthcareEntitiesAction`.
+- Added property `fhir_bundle` to `AnalyzeHealthcareEntitiesResult`.
+- Added property `confidence_score` to `HealthcareRelation`.
+- Added enum `HealthcareDocumentType`.
 
 ### Breaking Changes